+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Changed TarantoolNullField class to singleton ([#195](https://github.com/tarantool/cartridge-java/pull/275))
 - Bump netty to 4.1.78 ([#280](https://github.com/tarantool/cartridge-java/issues/280))
+- Enable Short to Integer converters ([#282](https://github.com/tarantool/cartridge-java/issues/282))
 
 ### Security
 - Bump jackson-databind to 2.13.4 ([#279](https://github.com/tarantool/cartridge-java/pull/279))